Create c++ home energy
In this project we use the combination of Arduino functionalities to monitory battery voltage (energy state). There many scenarios in a home where non-essential loads may need to be switched off and priority given to other loads especially when the energy is insufficient.
This project uses functionalities of Arduino combined with a potential divider and relay circuits to monitor the battery state of energy storage. One room contains loads with highest priority and its connected directly to the battery. At full charge, all rooms are powered but as battery voltage, Arduino automatically switches off the first room, further drop in voltage results in switching off the next room(load), leaving enough energy for the priority loads in room A .
The project further comes with an LCD screen display located in the main room , a person can know the battery energy state at any time and the rooms that have been turned off .At critically low voltage, the LED at the screen flashes constantly and the buzzer beeps to prompt.